Create submission_handler.py
