+++ dataactbroker/fileRoutes.py
@@ -1,6 +1,5 @@
-from functools import wraps
 from datetime import datetime
-
+from functools import wraps
 from flask import request, g
 from sqlalchemy import desc
 from webargs import fields as webargs_fields, validate as webargs_validate
@@ -9,16 +8,19 @@ from webargs.flaskparser import parser as webargs_parser, use_kwargs
 from dataactbroker.handlers.fileHandler import (
     FileHandler, get_error_metrics, get_status, list_submissions as list_submissions_handler,
     narratives_for_submission, submission_report_url, update_narratives, list_certifications, file_history_url)
-from dataactcore.interfaces.function_bag import get_submission_stats, get_fabs_meta
-from dataactcore.models.lookups import FILE_TYPE_DICT
+from dataactbroker.handlers.submission_handler import delete_all_submission_data, get_submission_stats
+
 from dataactbroker.permissions import requires_login, requires_submission_perms
-from dataactcore.models.lookups import FILE_TYPE_DICT_LETTER, JOB_STATUS_DICT, PUBLISH_STATUS_DICT
+
+from dataactcore.interfaces.db import GlobalDB
+from dataactcore.interfaces.function_bag import get_fabs_meta
+
+from dataactcore.models.lookups import FILE_TYPE_DICT, FILE_TYPE_DICT_LETTER, PUBLISH_STATUS_DICT
+from dataactcore.models.jobModels import Submission, Job, CertifyHistory, SubmissionWindow
+
 from dataactcore.utils.jsonResponse import JsonResponse
 from dataactcore.utils.responseException import ResponseException
 from dataactcore.utils.statusCode import StatusCode
-from dataactcore.interfaces.db import GlobalDB
-from dataactcore.models.jobModels import (
-    Submission, SubmissionSubTierAffiliation, Job, CertifyHistory, SubmissionWindow)
 
 
 # Add the file submission route
@@ -373,29 +375,7 @@ def add_file_routes(app, create_credentials, is_local, server_path):
     def delete_submission(submission):
         """ Deletes all data associated with the specified submission
         NOTE: THERE IS NO WAY TO UNDO THIS """
-
-        if submission.publish_status_id != PUBLISH_STATUS_DICT['unpublished']:
-            return JsonResponse.error(ValueError("Submissions that have been certified cannot be deleted"),
-                                      StatusCode.CLIENT_ERROR)
-
-        sess = GlobalDB.db().session
-
-        # Check if the submission has any jobs that are currently running, if so, do not allow deletion
-        jobs = sess.query(Job).filter(Job.submission_id == submission.submission_id,
-                                      Job.job_status_id == JOB_STATUS_DICT['running']).all()
-
-        if jobs:
-            return JsonResponse.error(ValueError("Submissions with running jobs cannot be deleted"),
-                                      StatusCode.CLIENT_ERROR)
-
-        sess.query(SubmissionSubTierAffiliation).filter(
-            SubmissionSubTierAffiliation.submission_id == submission.submission_id).delete(
-                synchronize_session=False)
-        sess.query(Submission).filter(Submission.submission_id == submission.submission_id).delete(
-            synchronize_session=False)
-        sess.expire_all()
-
-        return JsonResponse.create(StatusCode.OK, {"message": "Success"})
+        return delete_all_submission_data(submission)
 
     @app.route("/v1/check_year_quarter/", methods=["GET"])
     @requires_login
